Updating again EmployeeUpate

+++ src/main/java/com/fujitsu/ph/tsup/employee/management/web/EmployeeUpdateController.java
@@ -28,7 +28,7 @@ public class EmployeeUpdateController {
         }
 
         EmployeeUpdateForm employeeUpdate = new EmployeeUpdateForm();
-        employeeUpdate.setId(7);
+        employeeUpdate.setId((long) 7);
         employeeUpdate.setNumber("220053693");
         employeeUpdate.setLastName("Iwarat");
         employeeUpdate.setFirstName("Harvey");
